# Added optional type system - complain if proven wrong