--- a/source/syntax/mul.rs
+++ b/source/syntax/mul.rs
@@ -1,3 +1,3 @@
-fn mul(x, y) {
+fn mul(x, y) -> {
     return 3 * x * y
 }