--- a/source/syntax/mul_add.rs
+++ b/source/syntax/mul_add.rs
@@ -1,3 +1,3 @@
-fn add_mul(x, y) {
+fn add_mul(x, y) -> {
     return x * y + 3
 }