--- a/source/syntax/pos_len.rs
+++ b/source/syntax/pos_len.rs
@@ -1,3 +1,3 @@
-fn len(p) {
+fn len(p) -> {
     return sqrt(p.x^2 + p.y^2)
 }