--- a/source/syntax/return_void.rs
+++ b/source/syntax/return_void.rs
@@ -1,6 +1,5 @@
 fn foo() -> {
     return = "hi!"
-    return
 }
 
 fn main() {