--- a/source/syntax/typeof.rs
+++ b/source/syntax/typeof.rs
@@ -6,7 +6,6 @@ fn typeof_return() -> {
     if typeof(return) == "number" {
         println("return is number OK")
     }
-    return
 }
 
 fn main() {