--- a/source/test.rs
+++ b/source/test.rs
@@ -1,6 +1,7 @@
+fn foo(a: bool, b) {
+    println(a)
+}
+
 fn main() {
-    v := (1, 0, 0)
-    u := (0, 1, 0)
-    x := v x u
-    println(x)
+    foo(true, false)
 }