--- a/source/typechk/add.rs
+++ b/source/typechk/add.rs
@@ -1,0 +1,7 @@
+fn foo() -> vec4 {
+    return (1, 2) + 4
+}
+
+fn main() {
+    println(foo())
+}