--- a/source/typechk/call.rs
+++ b/source/typechk/call.rs
@@ -1,0 +1,8 @@
+fn foo(x: 'return f64) -> f64 {
+    return x
+}
+
+fn main() {
+    y := false
+    println(foo(y))
+}