--- a/source/typechk/or_2.rs
+++ b/source/typechk/or_2.rs
@@ -1,6 +1,6 @@
 
 fn foo() -> bool {
-    return false or 5
+    return true or 5
 }
 
 fn main() {