--- a/source/typechk/pow_2.rs
+++ b/source/typechk/pow_2.rs
@@ -1,0 +1,8 @@
+
+fn foo() -> bool {
+    return false ^ false
+}
+
+fn main() {
+    println(foo())
+}