--- a/source/typechk/pow_3.rs
+++ b/source/typechk/pow_3.rs
@@ -1,0 +1,8 @@
+
+fn foo() -> bool {
+    return "hi" ^ false
+}
+
+fn main() {
+    println(foo())
+}