--- a/source/typechk/return_3.rs
+++ b/source/typechk/return_3.rs
@@ -1,0 +1,7 @@
+fn foo() -> f64 {
+    return "hi"
+}
+
+fn main() {
+    println(foo())
+}