--- a/source/typechk/return_5.rs
+++ b/source/typechk/return_5.rs
@@ -1,5 +1,5 @@
 fn foo() -> bool {
-    return {}
+    return [0]
 }
 
 fn main() {