--- a/source/typechk/return_6.rs
+++ b/source/typechk/return_6.rs
@@ -1,5 +1,5 @@
 fn foo() -> bool {
-    return {}
+    return [0; 2]
 }
 
 fn main() {