--- a/source/typechk/return_7.rs
+++ b/source/typechk/return_7.rs
@@ -1,5 +1,5 @@
 fn foo() -> bool {
-    return {}
+    return (0, 1)
 }
 
 fn main() {