--- a/source/typechk_weak/mat_expr.rs
+++ b/source/typechk_weak/mat_expr.rs
@@ -1,0 +1,13 @@
+
+// fn f() -> vec4 { return (1, 2) }
+
+f() = (1, 2)
+
+fn foo() -> f64 {
+    // `f` returns `any`, so the sum of expression gets `any`
+    return f() + 1
+}
+
+fn main() {
+    println(foo())
+}