--- a/src/ast.rs
+++ b/src/ast.rs
@@ -34,7 +34,7 @@ pub struct Function {
     pub file: Arc<String>,
     pub args: Vec<Arg>,
     pub block: Block,
-    pub returns: bool,
+    pub ret: Type,
     pub source_range: Range,
 }
 
@@ -53,7 +53,7 @@ impl Function {
         let mut args: Vec<Arg> = vec![];
         let mut block: Option<Block> = None;
         let mut expr: Option<Expression> = None;
-        let mut returns = false;
+        let mut ret = Type::Void;
         loop {
             if let Ok(range) = convert.end_node(node) {
                 convert.update(range);
@@ -65,9 +65,15 @@ impl Function {
                     convert, ignored) {
                 convert.update(range);
                 args.push(val);
-            } else if let Ok((range, val)) = convert.meta_bool("returns") {
+            } else if let Ok((range, _val)) = convert.meta_bool("returns") {
                 convert.update(range);
-                returns = val;
+                if let Type::Void = ret {
+                    ret = Type::Any;
+                }
+            } else if let Ok((range, val)) = Type::from_meta_data(
+                    "ret_type", convert, ignored) {
+                convert.update(range);
+                ret = val;
             } else if let Ok((range, val)) = Block::from_meta_data(
                 "block", convert, ignored) {
                 convert.update(range);
@@ -87,7 +93,7 @@ impl Function {
         let block = match expr {
             None => try!(block.ok_or(())),
             Some(expr) => {
-                returns = true;
+                ret = Type::Any;
                 let source_range = expr.source_range();
                 Block {
                     expressions: vec![Expression::Return(Box::new(expr))],
@@ -113,10 +119,12 @@ impl Function {
             file: file,
             args: args,
             block: block,
-            returns: returns,
+            ret: ret,
             source_range: convert.source(start).unwrap(),
         }))
     }
+
+    pub fn returns(&self) -> bool { self.ret != Type::Void }
 }
 
 #[derive(Debug, Clone)]