@bvssvni 2016-05-12
This is just very basic type checking, but merging because it seem to work reasonably well and the more advanced stuff needs tweaking over longer time.